#27 클리어 위치 판정 추가 및 씬 구조 정리
