+++ Assets/Prefabs/Monsters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ef44380de4d060b4c92326ed229d5268
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
