+++ Assets/Prefabs/Monsters/BabySpider.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e1f0292b19ace9c4ba458686bcb50a5c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
