+++ Assets/Prefabs/Monsters/BlueButterfly.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a82e91a2fc15ea44ca77d0d59fba10d0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
