+++ Assets/Prefabs/Monsters/RedAnt.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6a578ccc7f11f9740824e9eeb4cc7885
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
