+++ Assets/Scripts/Clear.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 287b12af50b1d7a499f4700cb0426bef
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
