+++ Assets/Scripts/Clear/ClearArea.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0176b5101255ff4438a05f53b4878c5a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
