+++ Assets/Scripts/Clear/ClearController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6af6e3f9c25bc5741a7f2429e81cec83
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
